+++ Assets/Scripts/AugmentedScript.cs
@@ -5,8 +5,8 @@ using TMPro;
 
 public class AugmentedScript : MonoBehaviour
 {
-    public float currentLatitude;
-    public float currentLongitude;
+    public double currentLatitude;
+    public double currentLongitude;
     public double latitudeDrone; //39.7545206f 39.7325029f
     public double longitudeDrone; //-8.8016551f -8.820128f
     public double distance;
